dhcpv6-ia: create assignment for unknown IA in rebind messages
Create assignment for unknown identity association in rebind messages which is
equivalent to the handling of identity association for solicit/request messages.
However don't consider identity associations in rebind messages as a request so
that addresses/prefixes which are not apropriate for the link are returned with
lifetimes of 0.
This aligns with the behavior of the ISC DHCPv6 server.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>