router: make RA flags configurable (FS#2019)
Replace the config option ra_management by the config options ra_flags
and ra_slaac. The latter allows to configure the autonomous-address
config flag used for stateless address configuration while the former
allows to configure the RA flags in the form of a list.
The list can hold the following values :
	managed-config
	other-config
	home-agent
	none

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>