[RFC] Improve information provided in router advertisments
The two main changes are that Prefix Informations are only sent for
prefixes the router is responsible for and that Route Informations are
sent for all prefixes.

This allows to send router advertisements on the wan interface to tell
hosts there about the routes available via the router. And it also improves
routing between two downstream networks.

Unresolved issues:

 - valid and preferred life time needs to be determined
 - only lightly tested

Signed-off-by: Uwe Kleine-König <uwe@kleine-koenig.org>