dhcpv6-ia: allow up to 64 bit wide hostid
Add dhcpv6_hostid_len config option which controls the number
of bits in the host identifier of dynamically assigned IPv6
addresses. The default is 12 bits which is also the minimum.
The maximum is the whole interface identifier, i.e. 64 bits.

Allow up to 64 bit wide hostid in static leases.

Fixes #84 and #27.

Signed-off-by: Mikael Magnusson <mikma@users.sourceforge.net>