router: skip warning when configured lifetime is 0
The configured lifetime is overridden to be 0 in certain cases where
advertising a positive lifetime is not sensible. A log warning is only
warranted if the user was not already aware of this condition and
erroneously set the lifetime to a positive value.

If the user acknowledges these conditions by manually configuring a 0
lifetime, we should not warn them of the override.