relay: respect dhcpv6_[na|pd] config option
Take care of the uci config options dhcpv6_na and dhcpv6_pd in DHCPv6
relay mode as well.

If dhcpv6_na is set to false, remove the DHCPV6_OPT_IA_NA from the
client solicit message prior to forward the message to the upstream
DHCPv6 server.

If dhcpv6_pd is set to false, remove the DHCPV6_OPT_IA_PD from the
client solicit message prior to forward the message to the upstream
DHCPv6 server.

In the end, it allows to have control over what kind of addresses are
passed to the requesting client. Kind of similar to the already existing
feature "replace DNS server send by upstream DHCPv6 server".

Signed-off-by: Mathias Kresin <dev@kresin.me>