odchpd: support custom DHCPv6 options
This adds support for custom DHCPv6 options, similar to what dnsmasq supports
via "--dhcp-option".  Currently, only one type of option data is supported, and
that is a hexadecimal string defining the exact content of the option. More
variants can always be added later.

In other words: the full range of option formats supported by dnsmasq is *not*
supported, but this feature is a bit of a niche feature, so this initial
support is hopefully still useful.

Signed-off-by: David Härdeman <david@hardeman.nu>